code split main pages
